+++ tests/__init__.py
@@ -3,11 +3,14 @@
 import os
 import shutil
 import tempfile
+import string
+import random
 
 import pytest
 
 from dtoolcore import generate_admin_metadata
 from dtool_irods.storagebroker import (
+    _mkdir,
     _rm_if_exists,
     IrodsStorageBroker,
 )
@@ -17,6 +20,13 @@ _HERE = os.path.dirname(__file__)
 TEST_SAMPLE_DATA = os.path.join(_HERE, "data")
 
 TEST_ZONE = "/jic_archive"
+
+
+def random_string(
+    size=9,
+    chars=string.ascii_uppercase + string.ascii_lowercase + string.digits
+):
+    return ''.join(random.choice(chars) for _ in range(size))
 
 
 @pytest.fixture
@@ -64,3 +74,16 @@ def tmp_uuid_and_uri(request):
         _rm_if_exists(irods_path)
 
     return (uuid, uri)
+
+
+@pytest.fixture
+def tmp_irods_collection_fixture(request):
+    collection = os.path.join(TEST_ZONE, random_string())
+
+    _mkdir(collection)
+
+    @request.addfinalizer
+    def teardown():
+        _rm_if_exists(collection)
+
+    return collection