+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- ``dtool_irods.storagebroker.IrodsStorageBroker.list_dataset_uris`` class method
+
 
 Changed
 ^^^^^^^